requirements-dev.txt
# These are required for developing the package (running the tests, building
# the documentation) but not necessarily required for _using_ it.
codecov
coverage
flake8
pytest
sphinx
twine
pre-commit
black
nbstripout
requests_mock
# These are dependencies of various sphinx extensions for documentation.
ipython
matplotlib
mistune <2.0.0  # temporary while sphinx sorts this out
myst-parser
numpydoc
sphinx-click
sphinx-copybutton
sphinxcontrib.openapi
sphinx_rtd_theme
